refactor: split shared.go into messages.go and render.go (D1)
messages.go — all tea.Msg type definitions and the URLProvider interface
render.go   — RenderPost, attachment helpers, listFooter, renderChatMessages
shared.go   — retains only filterAmbiguousKeyMsg, extractURLs, attachmentURLs

No behaviour changes.
